pyproject.toml
[project]
authors = []
dependencies = ["python-dotenv==1.0.1", "slack_bolt==1.20.1"]
description = ""
name = "src"
readme = "README.md"
requires-python = "== 3.12.5"
version = "0.1.0"


[tool.rye]
dev-dependencies = [
  "autoflake==2.3.1",
  "autopep8==2.1.0",
  "flake8-plugin-utils==1.3.3",
  "flake8-type-checking==2.9.0",
  "flake8==7.0.0",
  "hatch-vcs==0.4.0",
  "hypothesis==6.111.1",
  "ipykernel==6.29.4",
  "ipython==8.23.0",
  "isort==5.13.2",
  "jupyterlab-quarto==0.3.3",
  "jupyterlab==4.1.6",
  "jupyterthemes==0.20.0",
  "pylint==3.1.0",
  "ruff==0.4.1",
]

[project.scripts]
bot = "src.bot:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.ruff]
extend-include = ["*.ipynb"]
include = ["*.ipynb", "*.py", "*.pyi"]
indent-width = 2
line-length = 160
target-version = "py312"

[tool.ruff.lint]
dummy-variable-rgx = "(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
extend-select = [
  "A",
  "ANN001",
  "ANN201",
  "ANN202",
  "ANN204",
  "ANN205",
  "ANN206",
  "ANN401",
  "ASYNC",
  "B",
  "C4",
  "DTZ",
  "E101",
  "E4",
  "E7",
  "E703",
  "E711",
  "E712",
  "E721",
  "E722",
  "E731",
  "E9",
  "E902",
  "E999",
  "F",
  "FBT",
  "FURB",
  "I",
  "I001",
  "I002",
  "ISC",
  "NPY",
  "PD",
  "PERF",
  "PIE",
  "PL",
  "PTH",
  "Q",
  "RET",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "T",
  "TCH",
  "TID",
  "TRY",
  "UP007",
  "UP034",
  "W291",
  "W292",
  "W293",
  "W605",
]
fixable = ["ALL"]
ignore = ["ISC001", "PLR0912", "PLR0913", "W191"]
unfixable = ["B"]


[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-type-checking]
exempt-modules = ["typing", "typing_extensions"]
quote-annotations = true
runtime-evaluated-base-classes = ["pydantic.BaseModel", "sqlalchemy.orm.DeclarativeBase", "sqlmodel.SQLModel"]
runtime-evaluated-decorators = ["attrs.define", "pydantic.validate_call"]
strict = true


[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true
known-first-party = ["src"]
length-sort = true
